moving stt integration text to join it's bretheren. Also bumping slow test limit.
